Add loader.lisp, weekfrequency-adt.lisp, and weekfrequency-adt-test.lisp files
./run.sh output:
$ ./run.sh
This is SBCL 2.0.1.debian, an implementation of ANSI Common Lisp.
More information about SBCL is available at <http://www.sbcl.org/>.

SBCL is free software, provided as is, with absolutely no warranty.
It is mostly in the public domain; some portions are provided under
BSD-style licenses.  See the CREDITS and COPYING files in the
distribution for more information.

"Hello world!" To load "parachute":
  Load 1 ASDF system:
    parachute
; Loading "parachute"

        ？ WEEKFREQUENCY-TEST::WEEKFREQUENCY-ADT-TEST-SUITE
        ？   WEEKFREQUENCY-TEST::PARACHUTE-OK
  0.004 ✔     (is = x y)
  0.000 ✔     (is equal x y)
  0.010 ✔   WEEKFREQUENCY-TEST::PARACHUTE-OK
  0.015 ✔ WEEKFREQUENCY-TEST::WEEKFREQUENCY-ADT-TEST-SUITE

;; Summary:
Passed:     2
Failed:     0
Skipped:    0