Implement graph iterators for node and arrow traversal
$ ./test-graph.sh
This is SBCL 2.0.1.debian, an implementation of ANSI Common Lisp.
More information about SBCL is available at <http://www.sbcl.org/>.

SBCL is free software, provided as is, with absolutely no warranty.
It is mostly in the public domain; some portions are provided under
BSD-style licenses.  See the CREDITS and COPYING files in the
distribution for more information.
To load "parachute":
  Load 1 ASDF system:
    parachute
; Loading "parachute"

;; Summary:
Passed:    55
Failed:     0
Skipped:    0