feat: Add script to run my-file.lisp
This commit adds a new script, `2-run-my-file.sh`, which allows running the `my-file.lisp` Lisp file using SBCL. This script can be used to execute the Lisp code without manually typing the command.

The purpose of this change is to provide a convenient way to run the Lisp file and test its functionality.